Option to explicitly set the committer
It isn't always clear who is the person or app that initiated the
workflow (ie. `GITHUB_ACTOR`) and it may be desirable to explicitly set
who the committer is when the action is run.

This commit provides the option to set the committer using the
`GH_USER_NAME` and/or `GH_USER_EMAIL` environment variables. If neither
environment variable is set it will default to the original behaviour of
just using `GITHUB_ACTOR`.